+++ playgrounds/desktop/src/dropdown/index.vue
@@ -64,7 +64,7 @@
 </template>
 
 <script setup lang="ts">
-import type { DropdownExposed } from '@veltra/desktop/types'
+import type { DropdownExposed } from '@veltra/desktop'
 import { shallowRef } from 'vue'
 
 import CustomCard from '../card/custom-card.vue'