+++ playgrounds/desktop/src/tag/index.vue
@@ -37,7 +37,7 @@
 </template>
 
 <script lang="ts" setup>
-import type { ColorType } from '@veltra/desktop/types'
+import type { ColorType } from '@veltra/desktop'
 import { shallowRef } from 'vue'
 
 import CustomCard from '../card/custom-card.vue'