+++ playgrounds/desktop/src/tree/index.vue
@@ -47,7 +47,7 @@
 </template>
 
 <script lang="ts" setup>
-import type { TreeExposed } from '@veltra/desktop/types'
+import type { TreeExposed } from '@veltra/desktop'
 import { shallowReactive, shallowRef, watch } from 'vue'
 
 import CustomCard from '../card/custom-card.vue'